Installing new dependencies in the project and creating the 'IUserRepository' interface
Installing new dependencies in the project through Maven, adding 'spring-boot-data-jpa' to work with database communication, the 'lombok' package for easier creation of getter and setter methods for class properties, and the 'h2' package, which will serve as an in-memory database to store application data. Adding the necessary variables for the new dependencies in the 'application.properties' file. Finally, creating the 'IUserRepository' interface to represent a repository for data returned from the database and modifying the other classes to send request data to the database in the appropriate format and handle multiple responses for the request