Creating the 'tasks' route, installing 'bcrypt' for encryption, and creating a filter to validate user actions.
Creating the '/tasks' route and the 'tasks' package for the tasks that will be created by the user. Creating the 'TaskModel' classes to represent a task in the application, 'TaskController' to instantiate and save the task data in the database according to the request, and the 'ITaskRepository' interface to communicate with the database by sending and receiving data. Installing the 'bcrypt' package as a dependency to perform password encryption in the application. Finally, creating the 'filter' package and the 'FilterTaskAuth' class to authenticate the user when registering a task