Completing user validation, creating the functionalities to list registered notes and update a note.
Concluding the user validation for the 'tasks' route and only for this route. Updating the 'ITaskRepository' interface, creating the 'findByIdUser' method to retrieve tasks registered by a specific user and return these tasks in a list. Finally, implementing the functionalities to list tasks created by the user and to update a task, both in full and in part, by creating the 'utils' package and the 'Utils' class to assist in mapping properties according to the data sent in the request