Finishing the application, completing the validation of task creation, and preparing for deployment
Concluding the 'Todolist' application, completing the final steps of validation for creation and updates, checking whether the user wishing to modify the task is the same as the one who created it, and verifying whether the 'title' field provided has a maximum of 50 characters, both in the task creation and update stages. Creating the 'errors' package and the 'ExceptionHandlerController' class to manage and handle application errors. Lastly, installing the 'spring-boot-devtools' dependency to aid in the development process, automatically re-executing the application whenever a change is made, and preparing the project for deployment by generating the '.jar' file of the application and creating the Dockerfile.